refactor: configure totals transition inside its component
