+++ packages/bar/src/index.ts
@@ -2,7 +2,7 @@ export * from './Bar'
 export * from './BarCanvas'
 export * from './BarItem'
 export * from './BarTooltip'
-export * from './BarTotal'
+export * from './BarTotals'
 export * from './ResponsiveBar'
 export * from './ResponsiveBarCanvas'
 export * from './props'
